[Mod] seperate those monitor tables into monitor.py
